refactor(videos): extract create() into VideoGenerationService, unify provider routing (#835)
videos.controller.ts was a 1,370-line god-controller dominated by an 835-line
create() method. Extract the full generation workflow into a new
VideoGenerationService; the controller's create() is now a thin handler that
keeps all decorators/guards/interceptors and delegates. ~25 of the 27 deps move
off the controller.

Provider routing fix: the two divergent switch statements collapsed into a single
dispatchVideoGeneration() that routes via isFalDestination() (model-key.util),
matching the images-operations precedent. The second (multi-output) switch was
missing FAL_KLING_VIDEO_V3_PRO, FAL_VEO_3_1 and FAL_PIXVERSE_V6, so additional
outputs of those models silently fell through to Replicate. Routing now lives in
exactly one place, structurally eliminating the missing-key bug class.

Also:
- delete the no-op `Promise.all(... { where: {} })` block (did nothing).
- extract createVideoPlaceholderActivity() used by the primary, batch and
  non-batch paths (was triplicated activity-create + websocket-publish).
- extract resolveVideoModel() and ensureDeferredCredits() as private methods.

Behavior-preserving: routes, decorators, guards, status codes and JSON:API
response shape unchanged. videos.controller.spec.ts rewired to build the real
VideoGenerationService from the same mocks (no expectation changes) plus a new
regression test asserting every output of a multi-output FAL_VEO_3_1 /
FAL_KLING_VIDEO_V3_PRO / FAL_PIXVERSE_V6 request routes to FAL, never Replicate.

Closes #835
