+++ apps/server/api/src/collections/videos/videos.module.ts
@@ -22,6 +22,7 @@ import { VideosProvenanceController } from '@api/collections/videos/controllers/
 import { VideosRelationshipsController } from '@api/collections/videos/controllers/relationships/videos-relationships.controller';
 import { VideosUploadController } from '@api/collections/videos/controllers/upload/videos-upload.controller';
 import { VideosController } from '@api/collections/videos/controllers/videos.controller';
+import { VideoGenerationService } from '@api/collections/videos/services/video-generation.service';
 import { VideoMusicOrchestrationService } from '@api/collections/videos/services/video-music-orchestration.service';
 import { VideoProvenanceService } from '@api/collections/videos/services/video-provenance.service';
 import { VideosService } from '@api/collections/videos/services/videos.service';
@@ -96,6 +97,7 @@ import { forwardRef, Module } from '@nestjs/common';
     CreditsInterceptor,
     ModelRegistrationService,
     ModelsGuard,
+    VideoGenerationService,
     VideoMusicOrchestrationService,
     VideoProvenanceService,
     VideosService,
